+++ backend/routes/addEmployee.ts
@@ -92,33 +92,7 @@ router.post(
     } } 
 );
 
-router.get("/employeeList", verifyToken, async (req: Request, res: Response) => {
-  try {
-    const token = req.token;
-    if (!token || !process.env.SECRET) {
-      return res.status(401).json({ message: "Unauthorized: Missing token or secret" });
-    }
 
-    const decoded = jwt.verify(token, process.env.SECRET) as jwt.JwtPayload;
-    const companyId = decoded.data.id;
-
-    const employees = await prisma.employed_by.findMany({
-      where: { company_id: companyId },
-      include: { employee_users: true }
-    });
-
-    const transformedEmployees = employees.map(emp => ({
-      id: emp.employee_user_id,
-      first_name: emp.employee_users.first_name,
-      last_name: emp.employee_users.last_name,
-      role: emp.employee_users.role
-    }));
-
-    res.json(transformedEmployees);
-  } catch (error) {
-    res.status(500).json({ message: "Server error"});
-  } 
-});
 
 
 export default router;
